refactor: dto 내부 메서드 이름 수정 및 dto를 패키지 별로 정리
