+++ src/main/java/kitchenpos/dto/menuproduct/MenuProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.menuproduct;
 
 import kitchenpos.domain.MenuProduct;
