+++ src/main/java/kitchenpos/dto/orderlineitem/OrderLineItemCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.orderlineitem;
 
 public class OrderLineItemCreateRequest {
     private Long menuId;
