+++ src/main/java/kitchenpos/dto/product/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.product;
 
 import kitchenpos.domain.Product;
