+++ src/main/java/kitchenpos/ui/ProductRestController.java
@@ -1,8 +1,8 @@
 package kitchenpos.ui;
 
 import kitchenpos.application.ProductService;
-import kitchenpos.dto.ProductCreateRequest;
-import kitchenpos.dto.ProductResponse;
+import kitchenpos.dto.product.ProductCreateRequest;
+import kitchenpos.dto.product.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,6 +25,7 @@ public class ProductRestController {
     public ResponseEntity<ProductResponse> create(@RequestBody @Valid final ProductCreateRequest productCreateRequest) {
         final ProductResponse created = productService.create(productCreateRequest);
         final URI uri = URI.create("/api/products/" + created.getId());
+
         return ResponseEntity.created(uri)
                 .body(created);
     }
