+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.application;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.OrderTableChangeRequest;
-import kitchenpos.dto.OrderTableCreateRequest;
-import kitchenpos.dto.OrderTableResponse;
+import kitchenpos.dto.ordertable.OrderTableChangeRequest;
+import kitchenpos.dto.ordertable.OrderTableCreateRequest;
+import kitchenpos.dto.ordertable.OrderTableResponse;
 import kitchenpos.repository.OrderRepository;
 import kitchenpos.repository.OrderTableRepository;
 import kitchenpos.repository.TableGroupRepository;
